Consolidate Numeric
