--- app/src/DesignSystem/Numeric/Big.svelte
@@ -1,15 +1,0 @@
-<script>
-  export let style = null;
-</script>
-
-<style>
-  p {
-    font-family: "GT America Mono Bold";
-    font-size: 22px;
-    line-height: 27px;
-  }
-</style>
-
-<p {style}>
-  <slot />
-</p>
