--- app/src/DesignSystem/Numeric/Regular.svelte
@@ -1,15 +1,0 @@
-<script>
-  export let style = null;
-</script>
-
-<style>
-  p {
-    font-family: "GT America Mono Medium";
-    font-size: 16px;
-    line-height: 19px;
-  }
-</style>
-
-<p {style}>
-  <slot />
-</p>
