--- app/src/DesignSystem/Numeric/Small.svelte
@@ -1,15 +1,0 @@
-<script>
-  export let style = null;
-</script>
-
-<style>
-  p {
-    font-family: "GT America Mono Medium";
-    font-size: 12px;
-    line-height: 14px;
-  }
-</style>
-
-<p {style}>
-  <slot />
-</p>
