--- app/src/DesignSystem/Numeric/index.js
@@ -1,9 +1,0 @@
-import Regular from "./Regular.svelte";
-import Big from "./Big.svelte";
-import Small from "./Small.svelte";
-
-export default {
-  Regular,
-  Big,
-  Small
-};
