+++ app/src/DesignSystem/index.js
@@ -5,7 +5,7 @@ import Icon from "./Icon";
 import Input from "./Input.svelte";
 import CheckboxInput from "./CheckboxInput.svelte";
 import DirectoryInput from "./DirectoryInput.svelte";
-import Numeric from "./Numeric";
+import Numeric from "./Numeric.svelte";
 import Text from "./Text";
 import Title from "./Title";
 import Avatar from "./Avatar.svelte";
